Remote path example for Hugo is misleading
Using `/static/admin` in my Hugo environment places the file in `/static/admin/admin`, which is wrong.  (My environment is Hugo 0.41 and Git on Bitbucket.)  Using just plain `admin` works in my case, which tends to imply that this path is taken relative to `/static/` but that doesn't entirely explain the first behavior, so the code must be doing something slightly more complex.