Remote Admin path examples for Hugo are misleading
In my environment, using `/static/admin` puts index.html in `/static/admin/admin`, which is wrong.  Using just `admin` works.  My environment is Hugo 0.41 and Git on Bitbucket.  It looks like this path might be taken as relative to `/static/` in Hugo but that doesn't entirely explain the first behavior, so the code must be doing something a bit more complex.